* 세션 델리게이트를 인터페이스로 수정 * 코어를 인터페이스로 수정

+++ MultiSocketRUDP/MultiSocketRUDPServer/SessionSendContext.h
@@ -11,7 +11,7 @@
 
 struct SendPacketInfo;
 
-enum class IO_MODE : LONG
+enum class IO_MODE : unsigned int
 {
 	IO_NONE_SENDING = 0
 	, IO_SENDING
